refactor(samples): create documentation component
