+++ ReduxSimple.Samples/Counter/CounterPage.xaml.cs
@@ -1,11 +1,7 @@
-﻿using Microsoft.Toolkit.Uwp.UI.Controls;
-using ReduxSimple.Samples.Common;
-using ReduxSimple.Samples.Components;
+﻿using ReduxSimple.Samples.Components;
 using System;
 using System.Reactive.Linq;
-using System.Threading.Tasks;
 using Windows.UI.Xaml.Controls;
-using static ReduxSimple.Samples.Extensions.FileExtensions;
 
 namespace ReduxSimple.Samples.Counter
 {
@@ -41,26 +37,7 @@ namespace ReduxSimple.Samples.Counter
             HistoryComponent.Initialize(_store);
 
             // Initialize Documentation
-            InitializeDocumentationAsync();
-        }
-
-        private async void InitializeDocumentationAsync()
-        {
-            const string folder = "Counter";
-
-            await LoadMarkdownComponent(IntroductionMarkdownTextBlock, $"{folder}/Introduction.md");
-            await LoadMarkdownComponent(StateMarkdownTextBlock, $"{folder}/State.md");
-            await LoadMarkdownComponent(ActionsMarkdownTextBlock, $"{folder}/Actions.md");
-            await LoadMarkdownComponent(StoreMarkdownTextBlock, $"{folder}/Store.md");
-            await LoadMarkdownComponent(UserInterfaceMarkdownTextBlock, $"{folder}/UI.md");
-            await LoadMarkdownComponent(CodeBehindMarkdownTextBlock, $"{folder}/CodeBehind.md");
-            await LoadMarkdownComponent(DependenciesMarkdownTextBlock, $"{folder}/Dependencies.md");
-        }
-
-        private async Task LoadMarkdownComponent(MarkdownTextBlock markdownTextBlock, string filePath)
-        {
-            markdownTextBlock.SetRenderer<CodeMarkdownRenderer>();
-            markdownTextBlock.Text = await ReadFileAsync(filePath);
+            DocumentationComponent.LoadMarkdownFilesAsync("Counter");
         }
     }
 }
