+++ ReduxSimple.Samples/TicTacToe/TicTacToePage.xaml.cs
@@ -1,13 +1,9 @@
-﻿using Microsoft.Toolkit.Uwp.UI.Controls;
-using ReduxSimple.Samples.Common;
-using ReduxSimple.Samples.Extensions;
+﻿using ReduxSimple.Samples.Extensions;
 using System;
 using System.Linq;
 using System.Reactive.Linq;
-using System.Threading.Tasks;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-using static ReduxSimple.Samples.Extensions.FileExtensions;
 
 namespace ReduxSimple.Samples.TicTacToe
 {
@@ -108,26 +104,7 @@ namespace ReduxSimple.Samples.TicTacToe
             HistoryComponent.Initialize(_store);
 
             // Initialize Documentation
-            InitializeDocumentationAsync();
-        }
-        
-        private async void InitializeDocumentationAsync()
-        {
-            const string folder = "TicTacToe";
-
-            await LoadMarkdownComponent(IntroductionMarkdownTextBlock, $"{folder}/Introduction.md");
-            await LoadMarkdownComponent(StateMarkdownTextBlock, $"{folder}/State.md");
-            await LoadMarkdownComponent(ActionsMarkdownTextBlock, $"{folder}/Actions.md");
-            await LoadMarkdownComponent(StoreMarkdownTextBlock, $"{folder}/Store.md");
-            await LoadMarkdownComponent(UserInterfaceMarkdownTextBlock, $"{folder}/UI.md");
-            await LoadMarkdownComponent(CodeBehindMarkdownTextBlock, $"{folder}/CodeBehind.md");
-            await LoadMarkdownComponent(DependenciesMarkdownTextBlock, $"{folder}/Dependencies.md");
-        }
-
-        private async Task LoadMarkdownComponent(MarkdownTextBlock markdownTextBlock, string filePath)
-        {
-            markdownTextBlock.SetRenderer<CodeMarkdownRenderer>();
-            markdownTextBlock.Text = await ReadFileAsync(filePath);
+            DocumentationComponent.LoadMarkdownFilesAsync("TicTacToe");
         }
     }
 }
